Move files to running folder
