+++ test/unit-tests/test-runner/pre-test.spec.ts
@@ -1,4 +1,4 @@
-import { TestRunner } from "../../../core/test-runner";
+import { TestRunner } from "../../../core/running/test-runner";
 import { TestSet } from "../../../core/test-set";
 import { Expect, AsyncTest, TestCase, SpyOn, Setup, Teardown } from "../../../core/alsatian-core";
 import { TestFixtureBuilder } from "../../builders/test-fixture-builder";
