+++ tsconfig.json
@@ -76,16 +76,16 @@
         "./core/results/test-outcome.ts",
         "./core/results/test-results.ts",
         "./core/results/test-set-results.ts",
+        "./core/running/test-item.ts",
+        "./core/running/test-plan.ts",
+        "./core/running/test-runner.ts",
         "./core/spying/restorable-spy.ts",
         "./core/spying/spy-call.ts",
         "./core/spying/spy-on.ts",
         "./core/spying/spy.ts",
         "./core/test-fixture.ts",
-        "./core/test-item.ts",
         "./core/test-loader.ts",
         "./core/test-output.ts",
-        "./core/test-plan.ts",
-        "./core/test-runner.ts",
         "./core/test-set.ts",
         "./test/builders/output-stream-builder.ts",
         "./test/builders/test-builder.ts",
